requirements.txt
blinker==1.8.2
click==8.1.7
Deprecated==1.2.14
Flask==3.0.3
importlib-metadata==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
opentelemetry-api==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-flask==0.45b0
opentelemetry-instrumentation-wsgi==0.45b0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
packaging==24.0
python-dotenv==1.0.1
typing_extensions==4.11.0
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.18.1
